Cope with enviro non-plus boards, and missing PMS5003s
If the attached Hat is a non-plus enviro, it doesn't have the connector
for the PMS5003 particulate sensor. In this instance, even with the
`miniuart-bt` devicetree overlay applied, the will be no /dev/ttyAMA0,
meaning the pms5003 module will throw when initialising.

If the attached hat is an enviro+ but doesn't have the PMS5003 breakout
box plugged in, the pms5003 module will initialise but throw a serial
error when trying to read.

Both of these cases are expected and shouldn't cause a crash.